job-bot/requirements.txt
aiohttp==3.9.5
feedparser==6.0.11
supabase==2.10.0
pytrends==4.9.2
python-dotenv==1.0.1
